% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msCCA.R
\name{mCCA}
\alias{mCCA}
\title{mCCA: msCCA object}
\description{
mCCA: msCCA object

mCCA: msCCA object
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{mCCA$new()}}
\item \href{#method-beta_init_func}{\code{mCCA$beta_init_func()}}
\item \href{#method-direction_update_l1_single}{\code{mCCA$direction_update_l1_single()}}
\item \href{#method-direction_update_l0_single}{\code{mCCA$direction_update_l0_single()}}
\item \href{#method-direction_update_grid}{\code{mCCA$direction_update_grid()}}
\item \href{#method-direction_cv_grid}{\code{mCCA$direction_cv_grid()}}
\item \href{#method-prediction}{\code{mCCA$prediction()}}
\item \href{#method-plot_cv_func}{\code{mCCA$plot_cv_func()}}
\item \href{#method-direction_grow}{\code{mCCA$direction_grow()}}
\item \href{#method-direction_grow_lazy}{\code{mCCA$direction_grow_lazy()}}
\item \href{#method-clone}{\code{mCCA$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{mCCA$new(
  X,
  beta_init = NULL,
  norm_type = 1,
  cl = 0.1,
  eta = 0.2,
  eta_ratio = NULL,
  rho_tol = 1e-06,
  rho_maxit = 1000,
  l1proximal_tol = 1e-04,
  l1proximal_maxit = 1000,
  line_search = TRUE,
  line_maxit = 10,
  eta_low = NULL,
  eps = NULL,
  init_method = "rgcca",
  my_init_param = NULL,
  trace = TRUE,
  print_out = 50
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-beta_init_func"></a>}}
\if{latex}{\out{\hypertarget{method-beta_init_func}{}}}
\subsection{Method \code{beta_init_func()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{mCCA$beta_init_func(R)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-direction_update_l1_single"></a>}}
\if{latex}{\out{\hypertarget{method-direction_update_l1_single}{}}}
\subsection{Method \code{direction_update_l1_single()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{mCCA$direction_update_l1_single(
  X = NULL,
  R = NULL,
  beta_init = NULL,
  l1norm = NULL,
  l0norm = NULL,
  trace = F
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-direction_update_l0_single"></a>}}
\if{latex}{\out{\hypertarget{method-direction_update_l0_single}{}}}
\subsection{Method \code{direction_update_l0_single()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{mCCA$direction_update_l0_single(
  X = NULL,
  R = NULL,
  beta_init = NULL,
  l1norm = NULL,
  l0norm = NULL,
  trace = F
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-direction_update_grid"></a>}}
\if{latex}{\out{\hypertarget{method-direction_update_grid}{}}}
\subsection{Method \code{direction_update_grid()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{mCCA$direction_update_grid(
  X = NULL,
  R = NULL,
  beta_init = NULL,
  norm_grids = NULL,
  trace = F,
  update = T,
  silence = F
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-direction_cv_grid"></a>}}
\if{latex}{\out{\hypertarget{method-direction_cv_grid}{}}}
\subsection{Method \code{direction_cv_grid()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{mCCA$direction_cv_grid(
  nfolds = 10,
  foldid = NULL,
  seed = 2021,
  n.core = NULL,
  trace = F
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-prediction"></a>}}
\if{latex}{\out{\hypertarget{method-prediction}{}}}
\subsection{Method \code{prediction()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{mCCA$prediction(xlist.te)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plot_cv_func"></a>}}
\if{latex}{\out{\hypertarget{method-plot_cv_func}{}}}
\subsection{Method \code{plot_cv_func()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{mCCA$plot_cv_func()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-direction_grow"></a>}}
\if{latex}{\out{\hypertarget{method-direction_grow}{}}}
\subsection{Method \code{direction_grow()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{mCCA$direction_grow(model_idx = 1)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-direction_grow_lazy"></a>}}
\if{latex}{\out{\hypertarget{method-direction_grow_lazy}{}}}
\subsection{Method \code{direction_grow_lazy()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{mCCA$direction_grow_lazy(out, norm)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{mCCA$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
