% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msCCA.R
\name{msCCAl0}
\alias{msCCAl0}
\title{msCCAl0: wrapper function for msCCA with decaying l0 norm bound for multiple components}
\usage{
msCCAl0(
  xlist,
  ncomp,
  xlist.te = NULL,
  init_method = "soft-thr",
  nfolds = 7,
  foldid = NULL,
  norms = NULL,
  norm_grids = NULL,
  nlambda = 10,
  rho_maxit = 200,
  print_out = 100,
  rho_tol = 0.01,
  eps = NULL,
  mode = "full",
  trace = F
)
}
\description{
msCCAl0: wrapper function for msCCA with decaying l0 norm bound for multiple components
}
