% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msCCA.R
\name{msCCAl1func}
\alias{msCCAl1func}
\title{msCCAl1func: wrapper function for using msCCAl1 to estimate mCCA directions with a pre-determined component number}
\usage{
msCCAl1func(
  xlist,
  ncomp,
  xlist.te = NULL,
  init_method = "soft-thr",
  nfolds = 20,
  warm_up = 50,
  penalty.C = 2,
  foldid = NULL,
  l1norm_max = NULL,
  l1norm_min = NULL,
  eta_ratio = NULL,
  eta = NULL,
  eps = NULL,
  my_init_param = NULL,
  l1proximal_maxit = 10000,
  rho_tol = 0.001,
  rho_maxit = 5000,
  print_out = 100,
  early_stop = T,
  step_selection = "penalized",
  norm_varying_ridge = T,
  multi.core = "mclapply",
  seed = 2021,
  trace = T
)
}
\description{
msCCAl1func: wrapper function for using msCCAl1 to estimate mCCA directions with a pre-determined component number
}
