% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{my_sggca_cv}
\alias{my_sggca_cv}
\title{my_sggca_cv: sgcca tuning.}
\usage{
my_sggca_cv(
  xlist,
  sparsities_grids,
  connection = NULL,
  nfolds = NULL,
  foldid = NULL,
  n.core = NULL
)
}
\description{
my_sggca_cv: sgcca tuning.
}
