% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{rho_estimation}
\alias{rho_estimation}
\title{rho_estimation: estimates mCCA coefficients without deflation}
\usage{
rho_estimation(xlist, betas)
}
\description{
rho_estimation: estimates mCCA coefficients without deflation
}
