% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{rho_estimation_deflated}
\alias{rho_estimation_deflated}
\title{rho_estimation_deflated: estimates mCCA coefficients withdeflation}
\usage{
rho_estimation_deflated(xlist, betas)
}
\description{
rho_estimation_deflated: estimates mCCA coefficients withdeflation
}
