% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_sparseGCATGD.r
\name{sgca_init}
\alias{sgca_init}
\title{sgca initialization Inputs:
=======
A, B:       Pair of matrix to calculate generalized eigenspace (sigma, sigma0)
nu:         Parameter of ADMM, default set to 1
K:          nuclear norm constraint, equal to r
rho:     penalty parameter on the l_1 norm of the solution, scaled by
epsilon:    tolerance level for convergence in ADMM
maxiter:    maximum number of iterations in ADMM
trace:      if set to True will print all iterations 
Outputs:}
\usage{
sgca_init(A, B, rho, K, nu = 1, epsilon = 0.005, maxiter = 1000, trace = FALSE)
}
\description{
sgca initialization Inputs:
=======
A, B:       Pair of matrix to calculate generalized eigenspace (sigma, sigma0)
nu:         Parameter of ADMM, default set to 1
K:          nuclear norm constraint, equal to r
rho:     penalty parameter on the l_1 norm of the solution, scaled by
epsilon:    tolerance level for convergence in ADMM
maxiter:    maximum number of iterations in ADMM
trace:      if set to True will print all iterations 
Outputs:
}
