% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_sparseGCATGD.r
\name{sgca_tgd}
\alias{sgca_tgd}
\title{Function for Thredholded Gradient Descent
Inputs:
=======
A, B:         Pair of matrix to calculate generalized eigenspace (sigma, sigma0)
r:            Latent dimension
k:            sparsity
init:         Initialize estimator of generlized eigenspace, obtained from sgca_int. Need to be k-sparse
lambda:       penalty parameter of Lagrangian function f(L), default to be 0.01
convergence:  tolerance level for convergence in TGD
maxiter:      maximum number of iterations in TGD
plot:         if set to True will plot intermediate iterations, need to specify scale variable V (as shown in paper)
              default set to False
Outputs:
========
final_estimate:   final estimation of leading r sparse generalized eigenspace}
\usage{
sgca_tgd(
  A,
  B,
  r,
  init,
  k,
  lambda = 0.01,
  eta = 0.01,
  convergence = 0.001,
  maxiter = 10000,
  plot = FALSE
)
}
\description{
Function for Thredholded Gradient Descent
Inputs:
=======
A, B:         Pair of matrix to calculate generalized eigenspace (sigma, sigma0)
r:            Latent dimension
k:            sparsity
init:         Initialize estimator of generlized eigenspace, obtained from sgca_int. Need to be k-sparse
lambda:       penalty parameter of Lagrangian function f(L), default to be 0.01
convergence:  tolerance level for convergence in TGD
maxiter:      maximum number of iterations in TGD
plot:         if set to True will plot intermediate iterations, need to specify scale variable V (as shown in paper)
              default set to False
Outputs:
========
final_estimate:   final estimation of leading r sparse generalized eigenspace
}
