% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_sparseGCATGD.r
\name{sgcaTGD_single}
\alias{sgcaTGD_single}
\title{sgcaTGD_single
A: Sigma
B: Sigma0
r: rank
k: row sparsity levels}
\usage{
sgcaTGD_single(
  xlist,
  xagg,
  r,
  k,
  lambda = 0.01,
  eta = 0.01,
  convergence = 0.001,
  maxiter = 10000,
  plot = FALSE
)
}
\description{
sgcaTGD_single
A: Sigma
B: Sigma0
r: rank
k: row sparsity levels
}
