% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_sparseGCATGD.r
\name{sgcaTGD_wrapper}
\alias{sgcaTGD_wrapper}
\title{sgcaTGD_wrapper
A: Sigma
B: Sigma0
r: rank
ks: row sparsity levels}
\usage{
sgcaTGD_wrapper(
  xlist,
  xagg,
  r,
  foldid,
  ks = NULL,
  lambda = 0.01,
  eta = 0.01,
  convergence = 0.001,
  maxiter = 10000,
  plot = FALSE
)
}
\description{
sgcaTGD_wrapper
A: Sigma
B: Sigma0
r: rank
ks: row sparsity levels
}
