% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msCCA.R
\name{msCCAl1}
\alias{msCCAl1}
\title{msCCAl1: wrapper function for msCCA with decaying l1 norm bound for multiple components}
\usage{
msCCAl1(
  xlist,
  ncomp,
  xlist.te = NULL,
  init_method = "soft-thr",
  nfolds = 7,
  foldid = NULL,
  norms = NULL,
  norm_grids = NULL,
  nlambda = 10,
  rho_maxit = 200,
  print_out = 100,
  l1proximal_maxit = 10000,
  rho_tol = 0.01,
  eps = NULL,
  mode = "full",
  trace = F
)
}
\description{
msCCAl1: wrapper function for msCCA with decaying l1 norm bound for multiple components
}
